# docs/swagger.yaml
basePath: /
definitions:
  Blogs.CommentResponse:
    properties:
      _id:
        type: string
      comment:
        type: string
      commented_by_id:
        type: string
      created_at:
        type: string
      post_id:
        type: string
      reply_post_id:
        type: string
      updated_at:
        type: string
    type: object
  Blogs.CreateCommentRequest:
    properties:
      comment:
        type: string
      reply_post_id:
        type: string
    required:
    - comment
    type: object
  Blogs.CreatePostRequest:
    properties:
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
...
  Blogs.LikeResponse:
    properties:
      created_at:
        type: string
      liked_by_id:
        type: string
    type: object
  Blogs.PostResponse:
    properties:
      _id:
        type: string
      authors_id:
        type: string
      created_at:
        type: string
      likes:
        items:
          $ref: '#/definitions/Blogs.LikeResponse'
        type: array
      text:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  Blogs.UpdateCommentRequest:
    properties:
      comment:
        type: string
      reply_post_id:
        type: string
    type: object
  Blogs.UpdatePostRequest:
    properties:
      text:
        type: string
      title:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllCommentsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/Blogs.CommentResponse'
        type: array
    type: object
  handler.getAllPostsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/Blogs.PostResponse'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8086
info:
  contact: {}
  description: Blogs Server in Go using Gin framework
  title: Blogs Server API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get a list of all posts
      operationId: get-posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllPostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a new Post with provided data
      operationId: create-post
      parameters:
      - description: post data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Blogs.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new Post
      tags:
      - Posts
  /posts/{post_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing Post by ID
      operationId: delete-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an existing Post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Get a single Post by providing its ID
      operationId: get-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Blogs.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a Post by ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update an existing Post data with new data
      operationId: update-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: new post info
        in: body
        name: new_input
        required: true
        schema:
          $ref: '#/definitions/Blogs.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing Post data
      tags:
      - Posts
  /posts/{post_id}/comments:
    get:
      consumes:
      - application/json
      description: Get a list of all comments related to this post
      operationId: get-post-comments
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllCommentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Comments
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a new Comment with provided data
      operationId: comment-on-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: comment data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/Blogs.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new Comment
      tags:
      - Posts
  /posts/{post_id}/like:
    post:
      consumes:
      - application/json
      description: Like the post
      operationId: like-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Like the post
      tags:
      - Posts
  /posts/{post_id}/unlike:
    post:
      consumes:
      - application/json
      description: Unlike the post
      operationId: unlike-post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlike the post
      tags:
      - Posts
  /posts/comments/{comment_id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing Comment of the post by ID
      operationId: delete-post-comment
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an existing Comment of the post
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Get a single Comment of the post by providing its ID
      operationId: get-post-comment
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Blogs.CommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a Comment by ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update an existing Comment data of the post with new data
      operationId: update-post-comment
      parameters:
      - description: Comment ID
        in: path
        name: comment_id
        required: true
        type: string
      - description: new comment data
        in: body
        name: new_input
        required: true
        schema:
          $ref: '#/definitions/Blogs.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing Comment data
      tags:
      - Posts
  /posts/my:
    get:
      consumes:
      - application/json
      description: Get a list of all my posts
      operationId: get-my-posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllPostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All My Posts
      tags:
      - Posts
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"